requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.4.16
click==6.7
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
MySQL-python==1.2.5
networkx==2.1
numpy==1.14.3
openslide-python==1.1.1
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.1.0
six==1.11.0
subprocess32==3.5.0
Werkzeug==0.14.1
